Kafka2SplunkService/requirements.txt
kafka-python==1.4.7
splunk-sdk==1.6.11